+++ app/(auth)/_layout.tsx
@@ -3,8 +3,8 @@ import { Stack } from 'expo-router';
 
 function AuthLayout() {
   return (
-    <Stack>
-      <Stack.Screen name="index" options={{ headerShown: false }} />
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="index" />
     </Stack>
   );
 }
