+++ components/organisms/Guide/StartGuide.tsx
@@ -1,0 +1,14 @@
+import Typography from '@/components/atoms/Typography';
+import ContentView from '@/components/atoms/View/ContentView';
+
+function StartGuide() {
+  return (
+    <ContentView>
+      <Typography variant="body4">시작하기</Typography>
+      <Typography variant="heading3">JOIN에서 믿을 수 있는{'\n'}스터디 메이트를 찾아볼까요?</Typography>
+      <Typography variant="body1">매번 기운빠지게 하는 빌런 스터디원 때문에{'\n'}힘들었다면 JOIN 하세요!</Typography>
+    </ContentView>
+  );
+}
+
+export default StartGuide;
